% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blake2b.R
\name{blake2b}
\alias{blake2b}
\title{Cryptographic hash of R object using 'blake2b'}
\usage{
blake2b(robj, N = 32, type = "string")
}
\arguments{
\item{robj}{Any R object supported by \code{serialize()} or \code{saveRDS()}}

\item{N}{size of hash in bytes. valid range [1, 64]. Default: 32.
Anything below 32 is discouraged.  Note that shorter hashes are 
not truncations of longer hashes.}

\item{type}{Return type: 'raw' or 'string'.  Default: 'string'}
}
\value{
cryptographic hash of the serialized object
}
\description{
BLAKE2b is a general-purpose cryptographic hash function -  
it is not suited for hashing passwords and deriving cryptographic keys from passwords.
}
\details{
To derive cryptograph keys from passwords use \code{argon2()}
}
\examples{
blake2b(mtcars)
}
