% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blake2b.R
\name{blake2b_raw}
\alias{blake2b_raw}
\title{Calculate cryptographic hash or raw vector or string using 'blake2b'}
\usage{
blake2b_raw(x, N = 32, type = "string")
}
\arguments{
\item{x}{raw vector or string}

\item{N}{size of hash in bytes. valid range [1, 64]. Default: 32.
Anything below 32 is discouraged.  Note that shorter hashes are 
not truncations of longer hashes.}

\item{type}{Return type: 'raw' or 'string'.  Default: 'string'}
}
\value{
cryptographic hash of the raw vector or string
}
\description{
Calculate cryptographic hash or raw vector or string using 'blake2b'
}
\examples{
blake2b_raw("apple")
}
