% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key-exchange.R
\name{create_public_key}
\alias{create_public_key}
\title{Create a public key from your secret key for public key cryptography.}
\usage{
create_public_key(your_secret_key, type = "string")
}
\arguments{
\item{your_secret_key}{Your secret key. Can be a character string, a 32-byte raw vector
or a 64-character hex string (encoding 32 bytes). When a shorter character string 
is given, a 32-byte key is derived using the Argon2 algorithm.}

\item{type}{Should the data be returned as raw bytes? Default: 'string'. 
Possible values 'string' or 'raw'}
}
\value{
32-byte public key
}
\description{
Use your secret key and public key with \code{create_shared_key()} to 
exchange keys over an insecure channel (i.e. public-key cryptography)
}
\examples{
your_secret_key <- argon2('hello')
create_public_key(your_secret_key)
}
