% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{decrypt_obj}
\alias{decrypt_obj}
\title{Read an encrypted R object from file}
\usage{
decrypt_obj(filename, key)
}
\arguments{
\item{filename}{source filename containing an encrypted R object}

\item{key}{The encryption key. Can be a character string, a 32-byte raw vector
or a 64-character hex string (encoding 32 bytes). When a shorter character string 
is given, a 32-byte key is derived using the Argon2 algorithm and 
a default, fixed code (note: this is insecure).  
If this argument is not explicitly set by the user 
when calling the function, an attempt is made to fetch \code{'MONOCYPHER_KEY'} from
the session global options. 
It is recommended that a key be created external to this function call.}
}
\value{
None
}
\description{
Read an encrypted R object from file
}
\examples{
file <- tempfile()
encrypt_obj(head(mtcars), file, key = "hello")
decrypt_obj(file, key = "hello")
}
