% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isaac.R
\name{isaac}
\alias{isaac}
\title{Use ISAAC to generate cryptographically secure pseudorandom numbers}
\usage{
isaac(n)
}
\arguments{
\item{n}{number of random bytes to return. Acceptable range
[1, 1024]}
}
\value{
a raw vector of random bytes of the requested length
}
\description{
See \url{https://en.wikipedia.org/wiki/ISAAC_(cipher)} and 
\url{https://burtleburtle.net/bob/rand/isaacafa.html} for more information
on this random number generator.
}
\examples{
isaac(32)
}
