% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{encrypt}
\alias{encrypt}
\title{Save an encrypted R object to file}
\usage{
encrypt(robj, filename, key = getOption("MONOCYPHER_KEY", default = NULL))
}
\arguments{
\item{robj}{R object}

\item{filename}{destination filename for encrypted R object}

\item{key}{The encryption key. Can be a character string, a 32-byte raw vector
or a 64-character hex string (encoding 32 bytes). When a shorter character string 
is given, a 32-byte key is derived using the Argon2 algorithm and 
a default, fixed code (note: this is insecure).  
If this argument is not explicitly set by the user 
when calling the function, an attempt is made to fetch \code{'MONOCYPHER_KEY'} from
the session global options. 
It is recommended that a key be created external to this function call.}
}
\value{
None
}
\description{
Save an encrypted R object to file
}
\examples{
file <- tempfile()
encrypt(head(mtcars), file, key = "hello")
decrypt(file, key = "hello")
}
